+++ factura_electronica/hooks.py
@@ -97,7 +97,7 @@ doc_events = {
     "Sales Invoice": {
         "on_submit": ["factura_electronica.api.pos_calculations",
                     "factura_electronica.api.generate_access_number",
-                      "factura_electronica.api.buscar_nombre"],
+                      "factura_electronica.buscar_nit.buscar_nombre"],
         "after_insert": "factura_electronica.api.generate_access_number",
     },
     "Purchase Invoice": {
